requirements.txt
anthropic>=0.39.0
openai>=1.50.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.7.0
thefuzz[speedup]>=0.22.0
Jinja2>=3.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
respx>=0.22.0
ruff>=0.6.0
mypy>=1.11.0
PyYAML>=6.0.2
playwright>=1.52.0
